# openapi.yaml
openapi: 3.0.3
info:
  title: Financial Records API
  description: |
    REST API for managing financial records with role-based access control,
    JWT authentication, and analytics dashboards.

    ## Authentication
    All endpoints except `/api/auth/register` and `/api/auth/login` require a
    Bearer JWT token in the `Authorization` header.

    ## Response Envelope
    All successful responses are wrapped in a standard envelope:
    ```json
    { "success": true, "data": <payload> }
    ```
    All error responses use:
    ```json
    { "success": false, "error": { "code": "<ERROR_CODE>", "message": "<description>" } }
    ```

    ## Error Codes
    | HTTP Status | Error Code        | Description                          |
    |-------------|-------------------|--------------------------------------|
    | 400         | BAD_REQUEST       | Validation error / bad input         |
    | 400         | FOREIGN_KEY_ERROR | Related record not found (P2003)     |
    | 401         | UNAUTHORIZED      | Missing or invalid JWT token         |
    | 403         | FORBIDDEN         | Insufficient role permissions        |
    | 404         | NOT_FOUND         | Resource not found                   |
    | 404         | RECORD_NOT_FOUND  | Prisma record not found (P2025)      |
    | 409         | CONFLICT          | Duplicate resource (e.g. email)      |
    | 409         | DUPLICATE_ENTRY   | Prisma unique constraint (P2002)     |
...

    ## Role Permissions
...
  version: "1.0.0"
...
servers:
...
security: []
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ───────────────────────────────────────────────────── Reusable schemas
  schemas:
...
    Role:
      type: string
      enum: [ADMIN, ANALYST, VIEWER]
    Status:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    ErrorEnvelope:
      type: object
      description: All error responses use this standardized envelope.
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: BAD_REQUEST
            message:
              type: string
              description: Human-readable error message
              example: Validation failed

    # ── Auth DTOs ──
...
    SafeUser:
      type: object
      description: User object returned by register and user management (explicit select, no password or updatedAt)
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        role:
          $ref: "#/components/schemas/Role"
        status:
          $ref: "#/components/schemas/Status"
        createdAt:
          type: string
          format: date-time
          example: "2026-04-01T12:00:00.000Z"
...
    CreateUserRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
          example: Alice Smith
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          minLength: 6
          example: password123
        role:
          $ref: "#/components/schemas/Role"
    UpdateUserRequest:
      type: object
      description: All fields optional. Password cannot be updated via this endpoint.
      properties:
        name:
          type: string
          minLength: 2
          example: Alice Johnson
        email:
          type: string
          format: email
          example: alice.new@example.com
        role:
          $ref: "#/components/schemas/Role"
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/Status"
    UserListData:
      type: object
      description: Inner data for list users response
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SafeUser"
        total:
          type: integer
          example: 5

    # ── Record schemas ──
...
paths:
...
  /users:
    post:
      tags: [Users]
      summary: Create user (Admin only)
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/SafeUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    get:
      tags: [Users]
      summary: List users (Admin only)
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Role"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Status"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/UserListData"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by id (Admin only)
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/SafeUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    patch:
      tags: [Users]
      summary: Update user (Admin only)
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/SafeUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    delete:
      tags: [Users]
      summary: Delete user (Admin only)
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        "204":
          description: User deleted (no content)
        "400":
          description: Cannot delete your own account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                error:
                  code: BAD_REQUEST
                  message: Cannot delete your own account
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{id}/status:
    patch:
      tags: [Users]
      summary: Activate or deactivate user (Admin only)
      operationId: updateUserStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/SafeUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─────────────────────────────────────────────────── Records
...